+++ rr/uri.js
@@ -40,7 +40,7 @@ export default class URI extends RR {
       target: TINYDNS.octalToChar(rdata.substring(16)),
       ttl: parseInt(ttl, 10),
       timestamp: ts,
-      location: loc !== '' && loc !== '\n' ? loc : '',
+      location: loc.trim() || '',
     })
   }