+++ test/ds.js
@@ -1,3 +1,4 @@
+import { describe } from 'node:test'
 import * as base from './base.js'
 
 import DS from '../rr/ds.js'