+++ test/sshfp.js
@@ -1,3 +1,4 @@
+import { describe } from 'node:test'
 import * as base from './base.js'
 
 import SSHFP from '../rr/sshfp.js'