+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Notable changes to this project are documented in this file.
 
 #### Unreleased
 
+### [1.4.0] - 2026-03-21
+
+#### Changed
+
+- feat(toBind): added to nsec3param, rrsig
+- feat(toTinydns): added to key, nsec3param, nxt, openpgpkey
+- feat(fromTinydns): added to cert, https, key, nsec, nsec3param, nxt, openpgpkey, smimea, svcb
+- fix(naptr): fixed flags calc in fromTinydns
+- fix(tinydns): more robust unpackDomainName
+- test: switch test runner to node --test
+
 ### [1.3.1] - 2026-03-13
 
 - ci: update workflow triggers
@@ -338,3 +349,4 @@ Notable changes to this project are documented in this file.
 [1.2.4]: https://github.com/NicTool/dns-resource-record/releases/tag/1.2.4
 [1.3.0]: https://github.com/NicTool/dns-resource-record/releases/tag/v1.3.0
 [1.3.1]: https://github.com/NicTool/dns-resource-record/releases/tag/v1.3.1
+[1.4.0]: https://github.com/NicTool/dns-resource-record/releases/tag/v1.4.0